PWGHF: add single muon analysis task
This task is derived from the DQ framework and it is used to extract the observables on single muons (pT, eta, sign of the muon, DCA,  etc) needed for the HF-muon analysis. It is indeed linked to HF since the final goal is to measure muons from charm decays and muons from beauty decays, separately from the information estracted with this task.